OK-746: Refaktoroitu järjestyskriteerin muokkauksen logiikka erilliseen hookkiin
