+++ src/app/haku/[oid]/henkilo/[hakemusOid]/components/jarjestyskriteeri-types.ts
@@ -1,0 +1,5 @@
+export type JarjestyskriteeriParams = {
+  tila: string;
+  arvo: string;
+  selite: string;
+};
